Retry most requests 5 times
We are seeing some intermittent 404 errors when making various requests
in our Happo CI jobs:

> StatusCodeError: 404 - "<html>\r\n<head><title>404 Not
> Found</title></head>\r\n<body bgcolor=\"white\">\r\n<center><h1>404
> Not
> Found</h1></center>\r\n<hr><center>nginx</center>\r\n</body>\r\n</html>\r\n"

These 404s happen with a variety of requests, and the error message
appears to be generated by nginx, which Happo does not use. My suspicion
is that this is something else in our infrastructure that is being
flaky. While I investigate the root cause, I think making Happo a bit
more fault tolerant would help reduce some pain from our engineers when
stuff like this crops up.

Since we've added exponential backoff and jitter to the retry mechanism,
this should be relatively safe to do across the board.

I left this off of fetchReport because we have some logic that relies on
404s from this request, so this will require a deeper change to get it
right.