Send payloads as multipart/formdata
We've been making some changes on the happo.io server side to avoid
reading large payloads into memory. One of the places where we had to
process large json was when payloads were posted from
RemoteBrowserTarget. The server has been updated to allow data to be
sent via multipart/formdata.

A new requirement from this new feature is that the client sends along a
hash of the payload contents. This is used on the server to optimize
certain things. When we're sending a `pages` config, we need to
randomize the hash (this was previously done server-side).