Skip git checkout when run-happo is in quick mode
The first time that Happo runs in CI, it currently runs `git checkout`
to check out the SHA. However, git should already be on that SHA since
it is the first run.

At Airbnb, I noticed that this step can take ~10 seconds, so skipping
over it when it is unnecessary should give us a little speed boost. It
is unclear to me exactly why this would take so long, so maybe all of
the time will just move over to the next git operation, but this seems
like a reasonable thing to do regardless.