Be smarter about checking out and npm installing
A recent commit (7780feb) made happo-ci skip doing a git checkout for
the current sha. This is good, as it makes things faster (a git checkout
can take some time in a large repo). But it made me aware of a flaw in
the logic around "QUICK". In most cases, CURRENT_SHA will be the same as
HEAD. But in some cases (as is the case when CI runs on a constructed
merge commit), CURRENT_SHA might actually be different from HEAD. In
this case we actually want to perform the git checkout.

By adjusting the "QUICK" logic to instead rev-parse the HEAD to see what
sha we're currently on and compare with the sha in question, we can keep
the speed-up from 7780feb yet still do the right thing when CURRENT_SHA
is not HEAD.