Don't run comparisons on default branch builds in Travis
For a default branch build, the happo-ci-travis script would run a
comparison between the current commit and the commit that travis last
knew about (that it had built). This led to failing builds for people
on some master/default branch builds because there would be a red happo
status (needs review). We can make this better by detecting when there
is a master/default branch build and only run a single report for that
commit.

Unfortunately, there's no way in Travis to know what the default branch,
so we'll have to default on `master` and then let people pass in a
different base branch using BASE_BRANCH if master isn't the one. This is
similar to what the happo-ci-circleci script does (although it depends
on it even more by looking up the PREVIOUS_SHA using a git-rev-parse).