Prevent duplicate happo-bundle entries in static packages
The happo bundles were added twice: once manually and once from the tmp
folder. We only need one of them, and I think it's easiest if it's just
coming from the tmp folder.

This commit will make snap-request caching less likely because of
changed timestamps. Caching is already a problem with static builds and
this is most likely not making things worse.