Switch to using pngjs when reading images
Jimp is a pretty big dependency, and we only use it for deep image
comparisons (which is a feature with few users). Switching to a leaner
library will make the library smaller in size. Since we only handle png
files, we can focus on a library that deals with png.

I choose pngjs over some other (png-js, upng-js) mainly because it
has support for piping a stream directly into a bitmap. The benefit of
this is that memory consumption will be a little lower (we won't read a
full image buffer into memory). Image diffing still requires an
in-memory read of the bitmap however. pngjs also has zero dependencies,
and it's all node.js code. Some of the alternatives provide native
extensions to C code (or something similar), and promise better
performance (on the surface). A while back I ran through a perf
comparison for reading an image between sharp (native) and Jimp (node)
and found virtually no difference. Image manipulation however, that's
where perf differences were noticeable. But we're not manipulating here,
just reading.