Retry fetching image for deep-compare
Instead of failing the run right away, we can allow a few retries before
we give up loading an image. I'm using the async-retry library here to
do the retry logic. This is a library used in other happo projects as
well.

We have some custom retry logic in makeRequest as well. I'm going to
refactor that to use the async-retry module as well in an upcoming
change.