Switch to async-retry for makeRequest
The retry logic here was all custom. Now that we have a dependency on
async-retry, we might as well remove the custom retry code.

The custom logic handled exponential backoff and jitter. I believe
async-retry handles both of these: backoff is handled by the `retry`
module (a dependency of async-retry) per default. And jitter is applied
by async-retry per default (it is off by default in `retry`). See
https://github.com/tim-kos/node-retry#retrytimeoutsoptions and
https://www.npmjs.com/package/async-retry#api

Apart from slightly different retry timings, this change will update the
console warning when a retry is made. Before, the next retry time would
be logged. We don't have that information now, so we have to adjust the
log a little.