Batch-process diffs when deep-comparing
Instead of trying to deep-compare all diffs at once, we do them
incrementally 10 at a time.

While I was writing the test for this change, I came across an error in
fetchPng that caused an unhandled promise rejection. The fix was to move
the `.on('error')` call above the call to `pipe`.

I thought about making the batch size configurable but decided it was
probably not necessary.