Log less in compareReports
We are running into an issue where our CI job fails when the number
of diffs to compare is really long. The error messages are either
never output, or confusing ("echo: write error: Resource temporarily
unavailable") so we don't have any strong leads on what exactly
could be causing it, and we have yet been able to reproduce it
locally.

One theory we have is that we are logging too much too quickly in
node which is causing the command to fail. A lot of this logging is
pretty unnecessary anyway so we are hoping to narrow down the
problem by logging less.