Batch logs in compareReports
We've been running into some issues where this function causes jobs
to fail because we run out of file descriptors. We believe we've
already solved this by logging less in 44637b04c9837, but out of an
abundance of caution, we want to add some batching for the remaining
logs here since the added complexity is low.