Add in-memory fallback for reading pngs
This is another attempt at resolving the issue we're seeing at airbnb
where one PNG image can't be read by pngjs. The theory here is that it
has to do with how pngjs handles streams. This issue describes the same
symptoms and goes on to explain how pngjs has a bug with how it handles
certain chunk sizes:
https://github.com/lukeapage/pngjs/issues/129

I tried the repro case described in the thread but wasn't able to
trigger the bad behavior:
  fs.createReadStream(path.join(__dirname, "in", "tbbn0g04.png"), { highWaterMark: 2 })
      .pipe(new PNG());

I then tried constructing a test case with different chunk sizes for the
requested images, but none would trigger the bad behavior. Using the
faulty image from airbnb itself didn't trigger things either, so my
theory is that this might be depending on some os specific thing.

Without reproing, it's hard to say if the fix I'm implementing here will
help. But, it most likely won't hurt --

- Try reading the png from a stream directly at the external image url
- If that fails, read the external image into memory, then feed it to
  pngjs

Hopefully this belt and suspenders approach will do the trick!