Add support for async reports
The idea of an async report is that the happo command can finish as soon
as the snap-requests have been initialized. The server will then
construct the report once all snap-requests are done.

So far, only RemoteBrowserTarget is aware of async reports, but going
forward the `happo run` command will know about it too (in some way).