Avoid throwing when there are no examples found
At Airbnb, we have a pretty complex setup where we have a bunch of
independent projects that are using Happo, and they pull from a shared
Happo configuration. Each component can decide which viewport sizes it
should be rendered in, which maps to a set of Happo targets. We have a
bunch of targets that map to the breakpoints we use in production. Most
components only use a couple of targets, and no projects currently use
all of the targets at once.

Currently, Happo will throw an error if one of the targets does not have
any tests for it. As a workaround, we generate stub tests for each
target. However, that comes with a bunch of overhead that we would be
able to avoid if Happo didn't have this restriction.

I think that throwing and failing the job might help some people
discover that they have a problem with their test setup earlier, but in
practice, I'm not sure that's much more useful than just getting an
empty report. For Airbnb's use-case at least, a much better tradeoff
would be to allow for the flexibility of some targets not having any
examples, and then just logging a warning and skipping over it.

There's a lot of indirection here, so it isn't immediately clear to me
what the early return value should be in this case, but it looks like
the results eventually gets picked up by constructReport, which expects
it to be an array. I can't help but point out that TypeScript would help
us out here.