Execute targets serially in domRunner generateScreenshots
We've been seeing some issues with our Happo jobs in CI. Occasionally,
they exit with status code 137, which seems to be how Kubernetes handles
processes that have exhausted all of the available memory. This always
seems to happen after "Generating screenshots in 6 targets..." is
logged. We are using the domRunner with the JSDOM prerendering.

I suspect that the call to executeTargetWithPrerender, which will
initialize a JSDOM environment and then execute our bundle in it, uses a
lot of memory. Since we are doing this 6 times in parallel, one for each
of the targets we have configured, this ends up using up too much
memory.

To resolve this, I think we need to dial back the parallelization here.
Since these tasks are not using separate threads or anything like that,
they are all sharing the same CPU core. And since these are all likely
CPU-bound tasks, it doesn't seem like we are getting much out of the
parallelization that Promise.all is providing us anyway. So, instead of
controlling the amount of concurrency here, I decided to just remove it
entirely.

As a followup, I think we should be able to further optimize this by
re-using the JSDOM environment across targets.