Send pre-rendered payloads to happo.io earlier
When we changed the pre-render behavior to serial instead of parallel,
we also introduced a perf regression for remote execution of the
pre-rendered payloads. Instead of sending the payload off to happo.io
servers as soon as pre-rendering was done, we would wait until all
payloads were ready. To improve this, we can kick off the remote job as
soon as possible, but still keep the parallel pre-render execution.