Speed up and make happo-ci more robust when HAPPO_IS_ASYNC
Support for the HAPPO_IS_ASYNC env variable was recently added to
happo.io. So far, it's only been affecting `happo run` and `happo
compare`, but I'm making it affect `happo-ci` as well. In async mode, we
don't need to check out the previous commit.

This should help resolve issues with running happo in github actions,
where (by default) the current sha is not the tip of the PR. Instead,
it's a merge commit to the current master. Some context here:
https://github.com/actions/checkout/issues/27