Add <ErrorBoundary> to React examples
By using an error boundary component, we can catch errors that occur
inside components and provide useful error logs.

https://reactjs.org/docs/error-boundaries.html

The default renderWrapper seemed like the best place to introduce the
boundary.

I went back and forth between rethrowing the error or simply render it
and let the screenshot show the error. I ended up with a combo: the
error is logged (console.error), and the stack is rendered as the
result. This means that render errors will stop breaking the test run,
instead you'll get a screenshot with the error details. This is similar
to what we already do for the storybook integration.

To better contextualize the error, I'm passing in the component and
variant name to the error boundary.