Find asset paths in picture source srcsets
We received a report of a confusing Happo report. Looking at the source
of the "after" version, the problematic part seemed to be related to a
`<picture>` element with nested `<source>` elements that used `srcset`.
After downloading the assets zip file for this report, we noticed that
the assets for these images were not included. I suspected that this was
related to our asset finding code, which we confirmed by pushing up a
change that included regular `<img>` tags with the same assets, which
ended up fixing the confusing report.

To fix this bug, we can include these elements in our srcset extraction
query selector. Since `<source>` is used for `<picture>`, `<video>`, and
`<audio>`, I scoped the selector to source elements that descend from
picture elements to avoid also grabbing audio and video files, which I
think we do not want in these bundles.