Update to jsdom v16
This update will make inlined css variables work (see e63d83c for a
description of the issue).

I had to update jest as well to get it on jsdom v16.

I'm seeing this deprecation warning now when running teststhat I aim to
get rid of in a follow-up commit:

  The node type SpreadProperty has been renamed to SpreadElement at
  Object.isSpreadProperty