Drop Node 8 from Travis test matrix
This version had its end of life on December 31, 2019. While it's
possinble that customers are still using it, there's no use for us
trying to get our jest tests to work on it (they currently fail because
the recently updated jest does not support <10).

Things will likely work on Node 8, but from this point there will be no
official support.