Upgrade to babel 7
After updating jsdom and jest to the latest versions, we started getting
deprecation warnings for certain babel-related issues. I'm taking the
opportunity here to upgrade the whole project to the latest @babel
packages.

At the time of the babel 6>7 release, there were a lot of issues from
Happo users trying to get babel, webpack and Happo to play nicely
together. I decided at the time to stay on babel 6 in the project, but
handle babel 7 as well. It was never pretty but it was one way we could
get as many projects working with Happo as possible.

Now I'm going to assume most people are on babel 7, and I'm okay with
scrapping babel 6 completely.