Update archiver from v3 to v5
I'm about to update a bunch of dependencies, patch and minor versions
only. Before I do so however, I want to single out a few major upgrades
that I feel are safe. The first being archiver, which according to
release notes should still work the same way we used it in v3:

https://github.com/archiverjs/node-archiver/releases/tag/5.0.0
https://github.com/archiverjs/node-archiver/releases/tag/4.0.0