Support webpack 5
Webpack version 5 was recently released and it's time we make sure to be
compatible.

At first look, I was going to drop support for webpack 3, but then I
decided to avoid that for now to prevent having to make (another) major
version bump (I just bumped to v6).

The transition was mostly smooth except for two things:

- some dependency is relying on `path` for something. Webpack 4 would
  automatically shim built-ins from Node, but now it's erroring. I added
  a no-op `fallback` for `path` and things seem to be working.

- The react-test had a webpack config that was partly incompatible with
  webpack 5. I had to mess around with publicPath for file-loader
  (including upgrading file-loader) to get things in a stable format.