Include project name in logs
At Airbnb, we are running Happo on a bunch of different projects in
parallel. Since the logs end up getting all mixed together, it is hard
to tell what is actually happening. I'm hoping we can make this better
by including the project name in the log lines if it exists.

I'm not totally in love with this approach, but anything cleaner would
require a deeper refactor of the way logging works in this project, and
I'm not really interested in doing that right now.