Move printDuration off of Logger class
This is essentially a private method that we don't want consumers of
this class to be calling. To make this more apparent, I am moving it off
of the class and passing in the required arguments at all of the
callsites.

This is in support of some logger refactoring I am about to do to make
it always log new lines at the end of every message.