Log start() and success()/fail() on separate lines
In https://github.com/happo/happo.io/pull/189 we made some improvments
to make the logs clearer when running Happo on multiple projects in
parallel. After landing this change, I noticed that since we leave
lines open with start() to be completed by success() or fail(), we
still get some logs that are a bit weird or confusing.

To make this better, I am changing start() to always log a newline,
and then teaching success() and fail() to include the start message if
it exists.

Now that the logger is getting a bit more complex in its output, I
decided that it would be helpful to have full examples of what it is
logging in the test suite, and Jest inline snapshots work well for
this. In order for the timings to be stable, we need to mock out the
timer, and Jest mock timers will mock out Date, but not perf_hooks, so
I am also switching to the slightly less granular Date.now at this
time. I don't think we really care about anything less than 1ms
anyway, so this should be just fine.