Upload assets for prerender runs separately
To reduce traffic towards s3 and avoid having to be throttled, we can
attempt to reuse asset packages between runs. This is already done for
static runs (e.g. storybook) but it's a little trickier to reuse things
between prerendered runs because assets might differ based on the size
of the viewport. Still, we can use a cache of known asset packages and
avoid the upload if the package is known.

I plan on pairing this change with an update to happo servers as well,
where we avoid the s3 upload if we come across a package that we've seen
and uploaded before. The cache used here is simply local to the cli
invocation, but a cache on happo servers can live longer than that.