Send fallbackShas in compare call to happo.io
If a comparison can't be made because the base/previous/before sha
doesn't have a report on happo.io, the server will do some work to
fallback to another base report, by looking in the git log for a
matching report. It will only do so for the main/default branch in the
git repo however. To support PRs going to a non-default branch without
requiring the happo.io server to know anything about the branch, we can
send an ordered list of shas that are well suited to be the base.

The logic on happo.io servers is the following:
- if the base/previous/before sha has a report, use it as the base
- if not, look at the list of fallback shas and pick the first one with
  a matching report
- if we still don't have a base report, traverse the git log and look
  for a match.