Allow fallback sha git command to fail
Right now, happo-ci-github-actions fails to execute in some
environments. This is probably becase it uses a shallow clone of the git
repo. To make the CI script work better in these scenarios, we make the
fallback-sha fetching an add-on instead of a requirement (allowing the
git log command to fail).

Fixes #204