Add HAPPO_SKIP_START_JOB to happo-ci script
At Airbnb, we are using Happo with orchestration. When we originally
started using this, we had a race condition between the orchestration
API call and when the projects would start running Happo. In this
scenario, we needed to always run the start-job command, which ensures
that a job is created for these, and if orchestration is used then it is
smart enough to combine the separate start-job calls into a single
orchestrated job.

More recently, we've observed some database contention partly due to our
scale. In my investigation, I realized that we have since resolved this
race condition and now have a guarantee that the orchestrate endpoint
will be called before Happo is run on any project in the PR. This means
that we can save a bunch of work by skipping this start-job command.

To make this skipping possible, I am adding a check for a
HAPPO_SKIP_START_JOB environment variable. This will allow us to switch
this behavior off and on if necessary.