Don't expand git shas that are already expanded
As one of the first steps in the happo-ci script, we expand the
CURRENT_SHA and PREVIOUS_SHA variables to normalize them. In
a github-actions environment, this can fail when the code has been
checked out using a shallow clone. To prevent that failure, we can
rev-parse only when needed.