Make HAPPO_IS_ASYNC=true the default
When the HAPPO_IS_ASYNC environment variable is set to "true", the
happo-ci script won't check out the previous commit and create a report.
Instead, we rely on that report (or a fallback) to already exist.

Async runs are faster and frees up valuable CI runtime since they don't
wait for the screenshots to be ready before exiting.

This is a breaking change for anyone using the happo-ci script directly,
the happo-ci-travis script, and the happo-ci-circleci script. The
happo-ci-github-actions script was already using async mode (it was
added later on and had async from the start).