Make dependency on Node >=10 explicit
We were already only testing the library in Node 10 and up, but there
was no explicit mention about the library only being support on Node 10
and later. Adding the `engines` section to package.json gets this
resolved.

While I was in this file, I ran it through prettier to get better
formatting.