Allow custom webpack
I'm debugging an issue with using Next@11 and webpack 5. Next.js has its
own bundled webpack, and if we use our own webpack we run into issues
with an `instanceof` check in webpack 5 (the class to check against is a
copy of the class we have, not the identical class).

To resolve this, I'm thinking that it might make sense to pass a webpack
instance directly from .happo.js. In the next.js case, this seems to
help get passed that instanceof check at least.