Ensure that remoteRunner creates consistent hashes
An important performance optimization is that we only store new static
packages and asset packages if they need to be stored. To aid in that,
we generate a hash from the contents of the package. With the static
package, we want the hash to be the same if the files stay the same. To
make that happen, we can inject a specific datetime for each zip file
entry. That way the zip package is deterministically produced.