Switch to using a temporary file when creating static packages
Reading a large zip file in memory isn't the best idea. Instead, we can
use streams. I thought about using a passthrough stream to avoid the
temporary file completely, but since we need the stream to create both
the md5 hash and the upload, I decided it was easier to just use a
temporary file. It's easier to debug as well since people can look for
happo-static-*.zip in their os.tmpdir() foler.