Fix staticPlugin check
The truthiness if staticPlugin is checked to determine if Happo should
use the remoteRunner, but since this is always assigned to an object
now, the logic was broken and Happo would never end up using the
domRunner. I believe we can fix this by being a little more careful
about how we assign the staticPlugin variable.