Check if assets need to be uploaded before uploading
The happo.io API soon has a new endpoint (added in 10.1.0) that we can
use to check if an asset/static package has already been uploaded or
not. We'll make another quick check right before the upload to check if
we can avoid it.

This will add a little bit of load to the server in the cold "caches"
case. But with time, this request will save us processing power.