Fix assertion in react-test.js
The react-test was failing as a result of applying fdc4f40. The
assertion was actually incorrect, and fdc4f40 definitely fixes a bug
with how assets from `publicFolders` were archived (shouldn't have
public folder path as part of name). This commit simply updates the test
to assert the right thing.