Export HAPPO_IS_ASYNC in happo-ci
I've had a report from a customer where the happo-ci flow is running
in async (HAPPO_IS_ASYNC=true) mode but the inner commands (happo run,
happo compare) are triggered in non-async mode. I think this is because
the HAPPO_IS_ASYNC variable isn't exported from the parent script.
Granted, the export isn't needed in most shell environments (my tests
show that it's enough to set it in the parent in bash and sh on Mac).
Exporting it makes the dependency clearer though, and is likely to fix
issues in certain shell environments.