Add happo-ci-azure-pipelines script
I was working on this script in a different repo yesterday.
https://dev.azure.com/trotzig/happo-demo-azure-full-page/_git/happo-demo-azure-full-page

It makes running Happo in an Azure DevOps pipeline a lot easier. The
trickiest part of the setup is getting the PREVIOUS_SHA. There are no
environment variables that tell us what commit on master that a pull
request is based on, and we only have a shallow git clone to work with.
So in order to run `git merge-base` we have to first fetch the source
code again, then extract local branch names before we can finally
resolve things.

On a build that isn't triggered for a pull request, we set the previous
sha equal to the current sha. This will trigger behavior in happo-ci
that avoids the comparison (only a single report will be created).