Allow individual pages to extend earlier report by sha
This is a new feature where you can specify certain pages to
inherit/extend from an earlier state instead of grabbing a new
screenshot from a url. This will help speed up certain test suites
because we can skip certain parts of the application.

The `extends` property is used instead of `url` for a page. It should
correspond to a commit hash/sha for a report that's been created before
(or is under construction right now).