Remove babel polyfill
The polyfill was added to the main entrypoint in 280f30f when supporting
a new `customizeWebpackConfig` option. At the time, we were supporting
Node versions 10 and later, but we've since moved on to support 12 and
later. Judging by this compatibility table, https://node.green/, if we
stop supporting v12 (which has been EOL) we can completely remove the
polyfill, which is what I'm doing now.

Fixes https://github.com/happo/happo.io/issues/239