Make requests time out after 120 seconds
When we moved from request-promise-native to node-fetch, we decreased
the timeout in makeRequest. This is causing some asset uploads to fail
at Airbnb. To fix this I'm making use of AbortController, which is
native in Node 15+ but available as an npm module in abort-controller
(recommended by node-fetch).

I've renamed the previous timeouts (which all had to do with retry
pauses) to make the API clearer.