Let snap-request POST finish before moving to next target
In 0be3322, we switched to serial execution of all targets. This helped
make things more stable and prevented contention on cpu and memory on
CI machines. In dd4d534103, I reverted some of this behavior to make
sure that we make the POST to snap-requests as soon as possible. I'm
pretty sure this has a bad effect on performance when there are large
payloads and targets. All POST requests are made slower which makes us
run into timeout issues and sockets being hung up.

To improve things, we can await the POST before we move to the next
target. We need to be careful and only do this in the async case,
because sync execution will poll and wait for results before continuing
to the next target.