Add data-happo-ignore to root element
This will cause it to be excluded from the rectangle calculation when
used with prerender:false.

When migrating one of my projects from prerender:true to prerender:false
I noticed that there was an extra rectangle around smaller examples.
Turns out the root element was causing this. In a prerender:true world,
we grab the innerHTML of this root node so it isn't present in rendering
on Happo workers.