Move scripts on happo debug page to bottom of document
I noticed that this page was broken for our internal test suite. By
moving the scripts to the bottom, just before the closing </body>, I
believe we can avoid some of the most common errors. In our case we were
expecting document.body to exist when the script was initialized.