Update archiver from v5 to v7
I am investigating a Jest test that verifies createStaticPackage creates
a deterministic hash when the file contents do not change, and it seems
to be flaky. I suspect that it might be using a timestamp in the hash
creation which would be a real issue we want to fix, but in the meantime
I figured we might as well update this package to the latest version.

They don't keep a useful changelog, but from what I can tell the
breaking changes only drop support for old node versions that we do not
support.

v6: https://github.com/archiverjs/node-archiver/compare/5.3.2...6.0.0
v7: https://github.com/archiverjs/node-archiver/compare/6.0.2...7.0.0